# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A marketplace API for buying and selling CS:GO skins",
        "title": "CS:GO Skin Marketplace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/marketplace/orders/{order_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific order by ID (only if owned by the user)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "marketplace"
                ],
                "summary": "Get order details",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - order not owned by user",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/marketplace/purchase": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Purchase a skin from the marketplace using user's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "marketplace"
                ],
                "summary": "Purchase a skin",
                "parameters": [
                    {
                        "description": "Purchase request",
                        "name": "purchase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.purchaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Purchase successful",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Skin not available",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Insufficient funds",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/marketplace/skins": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all skins available for purchase in the marketplace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "marketplace"
                ],
                "summary": "List available skins",
                "responses": {
                    "200": {
                        "description": "List of available skins",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Skin"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/marketplace/skins/mine": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all skins owned by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "marketplace"
                ],
                "summary": "List user's skins",
                "responses": {
                    "200": {
                        "description": "List of user's skins",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Skin"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/transactions/deposit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deposit a specified amount to the user's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Deposit money to balance",
                "parameters": [
                    {
                        "description": "Deposit request",
                        "name": "deposit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DepositRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deposit successful",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the transaction history for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction history",
                "responses": {
                    "200": {
                        "description": "Transaction history",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/withdraw": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Withdraw a specified amount from the user's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Withdraw money from balance",
                "parameters": [
                    {
                        "description": "Withdrawal request",
                        "name": "withdrawal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawal successful",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Insufficient funds",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Optional: include app error code",
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.purchaseRequest": {
            "type": "object",
            "required": [
                "skin_id"
            ],
            "properties": {
                "skin_id": {
                    "type": "string"
                }
            }
        },
        "models.DepositRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "models.Gun": {
            "type": "string",
            "enum": [
                "AK-47",
                "M4A4",
                "M4A1-S",
                "Desert Eagle",
                "USP-S",
                "Glock-18",
                "P250",
                "Tec-9",
                "CZ75-Auto",
                "AWP",
                "SSG 08",
                "SCAR-20",
                "G3SG1",
                "MP9",
                "MAC-10",
                "MP7",
                "P90",
                "UMP-45",
                "PP-Bizon",
                "Nova",
                "XM1014",
                "MAG-7",
                "Sawed-Off",
                "M249",
                "Negev",
                "Karambit",
                "Butterfly Knife",
                "M9 Bayonet",
                "Bayonet",
                "Flip Knife",
                "Gut Knife",
                "Huntsman Knife",
                "Shadow Daggers",
                "Falchion Knife",
                "Bowie Knife",
                "Navaja Knife",
                "Stiletto Knife",
                "Ursus Knife",
                "Nomad Knife",
                "Paracord Knife",
                "Survival Knife",
                "Classic Knife"
            ],
            "x-enum-varnames": [
                "AK47",
                "M4A4",
                "M4A1S",
                "DesertEagle",
                "USPS",
                "Glock18",
                "P250",
                "Tec9",
                "CZ75",
                "AWP",
                "SSG08",
                "SCAR20",
                "G3SG1",
                "MP9",
                "MAC10",
                "MP7",
                "P90",
                "UMP45",
                "PPBizon",
                "Nova",
                "XM1014",
                "MAG7",
                "SawedOff",
                "M249",
                "Negev",
                "Karambit",
                "Butterfly",
                "M9Bayonet",
                "Bayonet",
                "FlipKnife",
                "GutKnife",
                "Huntsman",
                "ShadowDaggers",
                "Falchion",
                "Bowie",
                "Navaja",
                "Stiletto",
                "Ursus",
                "Nomad",
                "Paracord",
                "Survival",
                "Classic"
            ]
        },
        "models.Order": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.OrderStatus"
                },
                "totalAmount": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.OrderStatus": {
            "type": "string",
            "enum": [
                "pending",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-comments": {
                "OrderStatusFailed": "Add this - when payment/processing fails"
            },
            "x-enum-varnames": [
                "OrderStatusPending",
                "OrderStatusCompleted",
                "OrderStatusFailed",
                "OrderStatusCancelled"
            ]
        },
        "models.Skin": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "condition": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "gun": {
                    "$ref": "#/definitions/models.Gun"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "rarity": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "balance_after": {
                    "type": "number"
                },
                "balance_before": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.TransactionType"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TransactionType": {
            "type": "string",
            "enum": [
                "withdraw",
                "deposit"
            ],
            "x-enum-varnames": [
                "Withdraw",
                "Deposit"
            ]
        },
...
        "models.WithdrawRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}